extract utils func
